getting users, template skeleton
