+++ main.go
@@ -3,10 +3,7 @@ package main
 import (
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-	"log"
-	"todo-demo/db"
+	"todo-demo/database"
 	"todo-demo/models"
 	"todo-demo/routers"
 )
@@ -24,14 +21,8 @@ func main() {
 	// #TODO: Finding a visible more pleasing formating of logs
 	e.Use(middleware.Logger())
 
-	// SQLite database connection
-	db, err := gorm.Open(sqlite.Open("todo-demo.db"), &gorm.Config{})
-	if err != nil {
-		log.Fatal("Error opening database:", err)
-	}
-
-	// Run migrations for all models - Comes from /db
-	migrations.Migrate(db)
+	// Initiate Database
+	database.InitiateDatabase()
 
 	// Registering the Routes
 	routes.RegisterRoutes(e)
