+++ templates/index.html
@@ -20,6 +20,9 @@
 
 <body class="bg-slate-800 text-white">
   This is it!
+  <button hx-get="/users" class='bg-slate-500 hover:bg-slate-900 p-2 rounded-lg transition'>
+    Click me
+  </button>
 </body>
 
 </html>
