+++ templates/todolist.html
@@ -1,0 +1,7 @@
+<ul>
+  {{range .todos}}
+  <li>
+    {{.Title}}
+  </li>
+  {{end}}
+</ul>
