+++ templates/userlist.html
@@ -1,0 +1,7 @@
+<ul>
+  {{range .users}}
+  <li>
+    {{.Age}}
+  </li>
+  {{end}}
+</ul>
